Fixed mismatch between schema.prisma and related migrations

+++ server/prisma/migrations/20231101001441_init/migration.sql
@@ -1,2 +1,2 @@
 -- AlterTable
-ALTER TABLE "workspaces" ADD COLUMN "similarityThreshold" REAL DEFAULT 0.25;
+ALTER TABLE "workspaces" ADD COLUMN "similarityThreshold" DOUBLE PRECISION DEFAULT 0.25;
